login

+++ routes/web.php
@@ -33,49 +33,59 @@ Route::get('/order/{code}', 'OrderController@show');
 Route::get('/contacts','ContactController@create')->name('contacts.create');
 Route::post('/contacts','ContactController@store')->name('contacts.store');
 
+//Login
+Route::get('login', [ 'as' => 'login', 'uses' => 'Auth\LoginController@getLogin']);
+Route::post('login', [ 'as' => 'login', 'uses' => 'Auth\LoginController@postLogin']);
+
+Route::get('register', 'Auth\RegisterController@getRegister')->name('register');
+Route::post('register', 'Auth\RegisterController@postRegister');
+
+Route::post('logout', 'Auth\LoginController@logout')->name('logout');
 
 /* -------------------------------------------------------------------------- */
 
 Route::prefix('admin')->group(function ()
 {
-	Route::get('/product', 'ProductController@index')->name('admin.products.index');
-	Route::get('/product/create', 'ProductController@create')->name('admin.products.create');
-	Route::post('/product', 'ProductController@store')->name('admin.products.store');
-	Route::get('/product/edit/{id}', 'ProductController@edit')->name('admin.products.edit');
-	Route::put('/product/{id}', 'ProductController@update')->name('admin.products.update');
-	Route::get('copy/product/{id}','ProductController@copy')->name('admin.products.copy');
-	Route::put('copy/product/{id}', 'ProductController@clone')->name('admin.products.clone');
-	Route::patch('copy/product/{id}', 'ProductController@clone')->name('admin.products.clone');
-	Route::delete('/product/{id}', 'ProductController@destroy')->name('admin.products.delete');
-	Route::get('/product-images/delete/{id}', 'ProductController@destroy_image');
+	Route::get('/product', 'ProductController@index')->name('admin.products.index')->middleware('auth');
+	Route::get('/product/create', 'ProductController@create')->name('admin.products.create')->middleware('auth');
+	Route::post('/product', 'ProductController@store')->name('admin.products.store')->middleware('auth');
+	Route::get('/product/edit/{id}', 'ProductController@edit')->name('admin.products.edit')->middleware('auth');
+	Route::put('/product/{id}', 'ProductController@update')->name('admin.products.update')->middleware('auth');
+	Route::get('copy/product/{id}','ProductController@copy')->name('admin.products.copy')->middleware('auth');
+	Route::put('copy/product/{id}', 'ProductController@clone')->name('admin.products.clone')->middleware('auth');
+	Route::patch('copy/product/{id}', 'ProductController@clone')->name('admin.products.clone')->middleware('auth');
+	Route::delete('/product/{id}', 'ProductController@destroy')->name('admin.products.delete')->middleware('auth');
+	Route::get('/product-images/delete/{id}', 'ProductController@destroy_image')->middleware('auth');
 
-	Route::get('/categories', 'CategoryController@index')->name('admin.categories.index');
-	Route::get('/categories/create', 'CategoryController@create')->name('admin.categories.create');
-	Route::post('/categories', 'CategoryController@store')->name('admin.categories.store');
-	Route::get('/categories/edit/{id}', 'CategoryController@edit')->name('admin.categories.edit');
-	Route::put('/categories/{id}', 'CategoryController@update')->name('admin.categories.update');
-	Route::delete('/categories/{id}', 'CategoryController@destroy')->name('admin.categories.delete');
+	Route::get('/categories', 'CategoryController@index')->name('admin.categories.index')->middleware('auth');
+	Route::get('/categories/create', 'CategoryController@create')->name('admin.categories.create')->middleware('auth');
+	Route::post('/categories', 'CategoryController@store')->name('admin.categories.store')->middleware('auth');
+	Route::get('/categories/edit/{id}', 'CategoryController@edit')->name('admin.categories.edit')->middleware('auth');
+	Route::put('/categories/{id}', 'CategoryController@update')->name('admin.categories.update')->middleware('auth');
+	Route::delete('/categories/{id}', 'CategoryController@destroy')->name('admin.categories.delete')->middleware('auth');
 
-	Route::get('/blogs', 'BlogController@index')->name('admin.blogs.index');
-	Route::get('/blogs/create', 'BlogController@create')->name('admin.blogs.create');
-	Route::post('/blogs', 'BlogController@store')->name('admin.blogs.store');
-	Route::get('/blogs/edit/{id}', 'BlogController@edit')->name('admin.blogs.edit');
-	Route::put('/blogs/{id}', 'BlogController@update')->name('admin.blogs.update');
-	Route::delete('/blogs/{id}', 'BlogController@destroy')->name('admin.blogs.delete');
+	Route::get('/blogs', 'BlogController@index')->name('admin.blogs.index')->middleware('auth');
+	Route::get('/blogs/create', 'BlogController@create')->name('admin.blogs.create')->middleware('auth');
+	Route::post('/blogs', 'BlogController@store')->name('admin.blogs.store')->middleware('auth');
+	Route::get('/blogs/edit/{id}', 'BlogController@edit')->name('admin.blogs.edit')->middleware('auth');
+	Route::put('/blogs/{id}', 'BlogController@update')->name('admin.blogs.update')->middleware('auth');
+	Route::delete('/blogs/{id}', 'BlogController@destroy')->name('admin.blogs.delete')->middleware('auth');
 
-	Route::get('/configurations','ConfigurationController@index')->name('admin.configurations.index');
-	Route::get('/configurations/edit/{id}','ConfigurationController@edit')->name('admin.configurations.edit');
-	Route::put('/configurations/{id}', 'ConfigurationController@update')->name('admin.configurations.update');
+	Route::get('/configurations','ConfigurationController@index')->name('admin.configurations.index')->middleware('auth');
+	Route::get('/configurations/edit/{id}','ConfigurationController@edit')->name('admin.configurations.edit')->middleware('auth');
+	Route::put('/configurations/{id}', 'ConfigurationController@update')->name('admin.configurations.update')->middleware('auth');
 
-	Route::get('/contacts','ContactController@index')->name('admin.contacts.index');
+	Route::get('/contacts','ContactController@index')->name('admin.contacts.index')->middleware('auth');
+	Route::get('/contacts/{id}','ContactController@show')->name('admin.contacts.show')->middleware('auth');
+	Route::delete('/contacts/{id}','ContactController@destroy')->name('admin.contacts.delete')->middleware('auth');
 
-	Route::get('/orders', 'OrderController@index')->name('admin.orders.index');
-	Route::get('/orders/status/{status}', 'OrderController@searchByStatus');
-	Route::get('/orders/search', 'OrderController@search');
-	Route::get('/orders/edit/{id}', 'OrderController@edit')->name('admin.orders.edit');
-	Route::get('/orders/{id}', 'OrderController@show')->name('admin.orders.show');
-	Route::put('/orders/{id}', 'OrderController@update')->name('admin.orders.update');
-	Route::delete('/orders/{id}', 'OrderController@destroy')->name('admin.orders.delete');
+	Route::get('/orders', 'OrderController@index')->name('admin.orders.index')->middleware('auth');
+	Route::get('/orders/status/{status}', 'OrderController@searchByStatus')->middleware('auth');
+	Route::get('/orders/search', 'OrderController@search')->middleware('auth');
+	Route::get('/orders/edit/{id}', 'OrderController@edit')->name('admin.orders.edit')->middleware('auth');
+	Route::get('/orders/{id}', 'OrderController@show')->name('admin.orders.show')->middleware('auth');
+	Route::put('/orders/{id}', 'OrderController@update')->name('admin.orders.update')->middleware('auth');
+	Route::delete('/orders/{id}', 'OrderController@destroy')->name('admin.orders.delete')->middleware('auth');
 });
